net: emac-thim: Fix driver crash on adb reboot
When the HW UP event does not come after the LPM suspend.
So the LPM Resume will not happen and if adb reboot is done.
Then the driver crash issue is seen.
To fix this we will not do the clean up in driver release.
Which are already done in driver suspend.