drivers:ioss: add condition to set the default value of wol only once
Currently the ioss driver sets the default wol setting whenever the
network interface is brought up. If the user changes the wol settings
and if there is any toggle in interface, the ioss overwrites the user
set value with default value. The fix ensures that default settings
are only made once after booting up.

Change-Id: I13397b63b6728880311a33231491a836f9477abf
Signed-off-by: Prachi Arora <quic_pracaror@quicinc.com>