data-eth: emac-ioss: Handle deletion of last filter properly
When dma filter is unique and last filter is to be deleted,
the filter is not deleted and considered as EMPTY commit.
Due to this the committed table is not cleared and shown
in eth-qos show. Handle this scenario by returning success
and deleting the filter properly.

Change-Id: Ie70c16271486cbeec7edb7997235d60cae79f2e5
Signed-off-by: Vatte Vishnuvardhan Reddy <quic_vishnuvr@quicinc.com>
(cherry picked from commit 07d53eb2c0e856bfa5f70b4196a901c31544f501)