drivers: ioss: Resolve incorrect IPA reconnect skip for Rx HW TCs
The QOS logic tries to avoid IPA reconnects as much as possible.
However, there is a special case in which a HW Rx QOS channels changes
to a SW one after a commit. Add this as an extra condition when trying
to check whether an IPA reconnect is needed or not.

Change-Id: If521643e12758701c7d5683343a4558e705decde
Signed-off-by: Mohd Ayaan Anwar <quic_mohdayaa@quicinc.com>