drivers: emac_ioss: Delete all filters during unique filter change
Right now if a commit causes the unique filter to change, some filters
escape deletion as the max filter range for different filters is
different. Fix this by setting all filters in the range [0,
max_filters_old] to be cleared. Also, add a debug log to print out the
filter table at the end of prepare_qos.

Change-Id: I29dcbd3250481f8b6f2f744b8e7cd30e62074983
Signed-off-by: Mohd Ayaan Anwar <quic_mohdayaa@quicinc.com>
Signed-off-by: Karthik Rudrapatna <quic_krudrapa@quicinc.com>