data-eth: ioss: Add IOSS recovery mechanism from ERROR state
Change the IOSS state to OFFLINE if it's in ERROR state while trying
to bringup IOSS online based on a netdev event. This will help provide a recovery mechanism in case IOSS went into ERROR state in the previous
netdev event cycle and avoid IOSS from the getting stuck in endless cycle of ERROR state.

Change-Id: Ie66cb5e2ba7a11734622661e2c9e323604bcc3de
Signed-off-by: Lakshit Tyagi <quic_ltyagi@quicinc.com>