data-eth: emac_ioss: Enhance show_qos information
The show_qos handler information isn't displayed when there are many
filters because the sysfs buffer length is limited to PAGE_SIZE.
And table pointer should not be accessd after kfree().

We are not ensuring Applied handler to display properly. We are
adding Applied handler to beginning so that QCMAP can get the Applied
handles when it parses the output.

Besides, fix a potential invalid pointer access.

Also, display DL information before UL as an enhancement since number
of DL filters is more limited and addressed corner case of filter
uniqueness.

Change-Id: I4e2f31f2335d79db8d7a53845f9bef530f139fee
Signed-off-by: Henry Yeh <quic_hanyyeh@quicinc.com>