data-eth: ioss: Add IOSS recovery mechanism from ERROR state
Change the IOSS state to OFFLINE if it's in ERROR state while trying
to bringup IOSS online based on a netdev event. This will help provide a recovery mechanism in case IOSS went into ERROR state in the previous
netdev event cycle and avoid IOSS from the getting stuck in endless cycle of ERROR state.

Change-Id: I74a13f4f614100919f72f65d943853cc41a35473
Signed-off-by: Shashank Shekhar <quic_sshe@quicinc.com>