drivers: emac_ioss: Fixing faulty pcp logging
The get_bit_pos API currently fails when multiple PCPs are mapped to a single queue. It is also not zero-indexed, which causes additional confusion. This change removes the use of that API and instead prints a hexadecimal value that can be used to determine the correct PCP-to-queue bitmap. Additionally, redundant logs indicating PFC enablement for PCP are removed.

Change-Id: Iec20b5123856020345b2aabb2a1edd2c56c4a5b0
Signed-off-by: Prachi Arora <quic_pracaror@quicinc.com>