drivers: emac_shim: Flush packets from Rx Queues when Rx stops
After DMA channel stop, flush the packet from the Rx Queues
destined to this DMA Rx Channel, when it is stopped. Clear it
so that on dma start because the packets that are received by
the MAC after the RxDMA is re-started are routed to the RxDMA
so no need to set it.

Change-Id: Iae015ea5d166e07df1941237f9ac4c4c4a85e91e
Signed-off-by: Sarosh Hasan <quic_sarohasa@quicinc.com>
(cherry picked from commit c824ec613891c74f7c3e4e4abb474e01e96ee500)