data-eth: emac_shim: handle thin driver fail scenarios
Handle scenarios when the thin driver fails to load properly.
When socket connection fails initially then driver init will not happen and if we try to do interface down then it will try to free resources which are not allocated in first place, causing device struck.
Scenarios handled are interface down/up and Driver suspend/resume.

Change-Id: I8907125e9cd0bd3fd9c9f6af0b0e208e4b0c5090
Signed-off-by: Sarosh Hasan <quic_sarohasa@quicinc.com>