data-eth: ioss: Add IOSS recovery mechanism from ERROR state
Change the IOSS state to OFFLINE if it's in ERROR state while trying
to bringup IOSS online based on a netdev event. This will help provide a
recovery mechanism in case IOSS went into ERROR state in the previous
netdev event cycle and avoid IOSS from the getting stuck in endless
cycle of ERROR state.

Change-Id: I59ee2f9f11554d3a524036cde3ad4f1e198cbf7c
Signed-off-by: Gaurav Nagar <quic_gauk@quicinc.com>